requirements.txt
Django==4.2
djangorestframework
django-ckeditor
googletrans==4.0.0-rc1
redis
django-redis
django-extensions
